Android: Started adding platform specific macro identifiers and code paths to support building JavaScriptCore on Android. Even though Android is a Linux derivative, Android doesn’t provide full access to the kernel headers and other very OS specific headers. Additionally, Android does not use glibc, and instead reimplemented a minimal C library replacement called Bionic which is neither POSIX or ANSI compliant (and thus missing things).
In this change set:

Platform.h:
Added the OS(ANDROID) support macro. Additionally, the OS(UNIX) marco will considered Android to be a UNIX.
TODO: I need to remove the #warning stuff (debugging compiles)

ThreadIdentifierDataPthreads.cpp:
PTHREADS_KEYS_MAX was already known not to be defined in Bionic according to the comment, but there was no OS(ANDROID) check for this which this fixes.

MacroAssemblerARM.cpp:
Android doesn't provide <asm/hwcap.h>, but cpu-features.c in the ndk in sources/android/cpufeatures has the following #defines referring to <asm/hwcap.h>, so I think it is safe to paste them in here with the OS(ANDROID) guard.

Assertions.cpp:
Android doesn’t have the header <execinfo.h> and it doesn’t have backtrace(). The code is commented out temporarily. This will break other platforms (not Windows though).
TODO: Add the OS(ANDROID) guards to disable backtrace() support just for Android and avoid breaking the other platforms.
TODO: Figure out if there is an alternative. (Some searching suggests it won’t be easy. There is a function that will dump traces to the log, but not return actual data.)