Android/CMake: These contain additional fixes/corrections/improvements to the build process via CMake for JavaScriptCore for Android.
BUILDING.txt:
The prior instructions had some fatal bugs. Building ICU as static only (no dynamic libraries) would result in failed linking for the jsc app. In trying to diagnose and resolve the problem, I have simplified the instructions so that U_DISABLE_RENAMING is no longer used. We now use the defaults which means ICU will name mangle their names with a version suffix. (The Source/WTF/wtf/CMakeLists.txt helps this change work.)

Source/JavaScriptCore/shell/PlatformAndroid.cmake:
This is a new file that instructs CMake to link against the ICU libraries when on Android. A line in WebKitMacros.cmake that I disabled earlier needed to be reactivated for this file to be read. (Presumably, I removed it because I didn’t have this file.) I had to add ICU_DATA_LIBRARIES to FindICU.cmake because it appears the core depends on it. I also thing the variables the BlackBerry counterpart uses are wrong.

Source/WTF/wtf/CMakeLists.txt:
I deleted the search path for icu headers which I had previously added. It didn’t belong here, and I think this directory shouldn’t even be in the source tree. We should always be pulling from the official headers that go with your version of the ICU libraries. Because of the weird name mangling ICU does, the headers are not compatible among versions and it is very dangerous to intermix the too. It is good this line is removed. I think this is also partly why we can now compile ICU with the default name mangling on.

Source/cmake/FindICU.cmake:
I added a FIND_LIBRARY for the udata library. The main icu library seems to depend on this via openCommonData in source/common/udata.cpp in 52.1. I don't know if this is a new development or just unnoticed by those dynamically linking everything and already have all the libraries. But its absence is noticed in static linking scenarios which I am currently developing under.

Source/cmake/OptionsAndroid.cmake:
Removed the explicit flag to disable name mangling for Android now that we can use the defaults.

Source/cmake/WebKitMacros.cmake:
Re-enabled the macro that includes PlatformAndroid.cmake. This was needed to provide the correct link flags for the jsc shell app which I was using as a linking test.