Android/CMake: Made modifications and changed instructions to build JavaScriptCore as a dynamic library on Android.
Building as a dynamic library has some advantages in this case.
1) When building the final Android NDK based JNI library for your app, you won’t have to bring in all the dependencies for JavaScriptCore (e.g. ICU, WTF).
2) Since JavaScriptCore is under the LGPL, statically linking code to it may be problematic for some people

In changing to dynamic linking, some new problems were found which this change set addresses. The most significant is that a initializeMainThreadPlatform() implementation (and a few others) must be defined or you get undefined symbol errors when trying to build JSCore. Looking at the Windows Phone 8 implementation, it is completely stubbed out. When I built jsc against the static library version of JavaScriptCore, I didn’t get the errors, which suggests that it is not in use by jsc. It could be that we don’t actually need these for raw JavaScriptCore (these functions are part of WTF).

BUILDING.txt:
-DSHARED_CORE=1 is added to the command line switches to build JavaScriptCore as a dynamic library

Source/JavaScriptCore/CMakeLists.txt:
Since I just went through needing to put the correct ICU link flags for jsc, I know I needed to make these available when building JSCore as a dynamic library.
Additionally, I disabled the CMake versioning features for Android because it is incompatible with Android/Java’s loadLibrary.

Source/WTF/wtf/CMakeLists.txt:
Added the data library based on previous lessons where I discovered I needed it for the icu core. This *might* be an issue for people if they don’t have the library on their system and they for whatever reason don’t have this dependency.

Source/WTF/wtf/PlatformAndroid.cmake:
This follows the convention already laid out to have a PlatformFOO.cmake file add the MainThreadFOO.cpp implementation.

Source/WTF/wtf/MainThreadAndroid.cpp:
This is a copy of the Windows phone implementation which is also a stub. I suspect this is not actually needed for JavaScriptCore, but is needed for WebKit. Anybody pushing harder to complete the WebKit port for Android will most likely need to complete this implementation.