Android test262: Added Mac client which can download the test262 log files from the Android device. (Mac client includes 3rd-party CocoaHttpServer)
There are multiple pieces that make this all work.
- The Android code uses Zeroconf to advertise itself (using the native 4.1 Zeroconf implementation)
- The Mac client uses Zeroconf (Bonjour) to discover the Android devices
- The Mac client can connect to the Android device once discovered
- The Mac client also runs a private http server which will be used for a way the Android device can upload the log files to
- When the Mac client connects to the device, it sends the port number of the private http server.
- The Android device (already knowing the address of the connected Mac client), takes the private http server port number and uses HttpURLConnection to upload the log file to the Mac client’s private http server

The code takes a lot from Apple and Android sample code.
Additionally, the 3rd party CocoaHTTPServer is used.

The client upload code was actually originally stubbed/prototyped using NSURLSession upload tasks. This is the reason the Content-Type is application/octet-stream (because this is what NSURLSession was setting when trying to send files instead of NSData). This code was designed so that the Mac client could also connect to an iOS test262 counterpart when I eventually finish writing that, hence the NSURLSession influence.

To initiate a download, double-click on the discovered service in the tableview. It will start downloading into the private http server’s docRoot directory. Lots of GUI is still needed, lots of fixes are needed (like changing the service type and download location), and lots of corner cases are not handled (like initiating multiple downloads at the same time or service resolve triggers multiple times because of ipv4 and ipv6).

I wanted to also add over-the-air logging, though this whole network communication thing turned out to be a much harder/more complicated than I had hoped. (I thought both the Cocoa and Java frameworks would have made this kind of thing much easier by now.)