test262: Started implemented network server and Bonjour advertisement. (Download log now works on Mac.) 3rd party SimpleThread added.
Most of the raw server is implemented in pure C/BSD sockets to try to maximize potential reuse. There is a function pointer entry that allows for the platform specific hook to do the final mile for uploading a file. In this case, NSURLSession is used.

The remaining network code is common to Mac/iOS, though the iOS project has not yet been updated to use all these changes, so it probably is broken.

SimpleThread was added to save me time because I wanted to write a common server and thus couldn’t use GCD. It is my very stripped down thread abstraction layer that pretty much does nothing except let you spawn and join threads. But that’s all I needed and it has a pthreads and Windows backend all written in C so it is very portable. SimpleThread is part of ALmixer of http://playcontrol.net/opensource/ALmixer.