test262: Android: Creating a new version of test262 that is more C-based for performance comparison to the Java-based version.
This is mostly a straight copy from the other version, except I started adding Test262HelperAndroid and CommonUtils.

We are investigating performance issues and see that the Dalvik garbage collector seems to be a big problem. Having a version in C to compare against will help us verify if this is really the issue.