test262: Android: Changed Zeroconf behavior to not advertise when the app is backgrounded. Also added more debug stuff.
There seems to be a bug in Android’s zeroconf implementation where if the process exits uncleanly (crash, adb install while running), the current advertisement persists and seems to be kept alive by the OS until a reboot. (Can’t connect failures fail to unregister from the zeroconf cache until the device is rebooted.) Removing the advertisement during backgrounding can help reduce the number of accidental cases where the app is reinstalled while it is still running. However, this doesn’t fix the fundamental problem.

Also discovered the hard way that double registration (same service on the same port results in double registrations). The onCreate() and onResume() code both start up zeroconf which resulted in a double registration which onPause() didn’t handle leaving the service still mysteriously running. The quick solution was to put a quick check in the NsdHelper.java itselff, though maybe a better solution is needed.

Some additional debugging stuff is added, so of which probably should be removed in the future.