test262: Android: Significant overhaul for C-centric version of Android test262.
Logic and infrastructure is now in place. GUI function callbacks are also in place, but are empty functions. The tests run, but the GUI needs to be wired into the new callbacks for the final step.

SocketServer is now shared among all the implementations. The socket and URL components were fixed to deal with subtle differences between Apple and Linux/Android. IPv6 for http URL generation is fixed. NetworkHelper.java is sort of a counterpart to the NetworkHelper.m in the Apple versions. It is mostly a launching pad to call the C-based SocketServer plus coordinate native platform Zeroconf implementations into the mix. The old LoggingConnection and other Java socket server stuff are now obsolete and should eventually be removed. (The HTTP stuff is still relevant.)

Logger bindings generated via SWIG were brought into help allow calling Logger as needed from the Java side. However, it turns out I didn’t need them at this time.